Student Mentorship/requirements.txt
mysql-connector-python==8.0.31
Pillow
tk
customtkinter
matplotlib
pywinstyles
CTkMessagebox
pandas
tktimepicker
tkcalendar
Reportlab